fix(identity): track paren depth alongside bracket depth in return-type scanners
A relational/shift operator inside a paren-wrapped non-type template
argument (e.g. std::enable_if_t<(sizeof(T) < 4), int>) permanently
corrupted _top_level_paren_spans's and _find_top_level_arrow's bracket
counter: bracket depth was tracked only at the outer scanning level, so
a '<' reached while already inside an unrelated template argument list
(bracket already nonzero) incremented the SAME counter with no way back
to zero, silently discarding the real parameter list, a trailing return
arrow, or an exception specification depending on where the operator
occurred.

Fixed by tracking bracket depth and paren depth together in one pass in
both scanners: a bare '<'/'>' is only ever counted as a bracket while
paren depth is also zero, since a non-type template argument containing
a relational/shift operator must be paren-wrapped per the grammar to
disambiguate it from the closing '>' -- so once such a paren group has
opened, any '<'/'>' inside it can only be that operator, never a genuine
template bracket.

Also split tests/test_entity_id_template_discriminators.py (grown past
the architecture gate's 1200-line test-file cap) into
tests/test_entity_id_return_type_discriminators.py for the
return_type()-parsing-specific tests; no test content changed by the
split itself.
